+++ detector/__init__.py
@@ -1,1 +1,6 @@
 # detector package
+from detector.rule_based import detect_prompt_injection
+from detector.hybrid import detect
+from detector.preprocessor import preprocess
+
+__all__ = ["detect_prompt_injection", "detect", "preprocess"]